% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aegean.R
\docType{data}
\name{aegean}
\alias{aegean}
\title{A SAR dataset describing invertebrates on islands in the Aegean Sea, Greece}
\format{A list with two elements. The first element contains the name of the
  dataset. The second element contains a data frame with 2 columns and 16
  rows. Each row contains the area of an island in the Galapagos (1st column)
  and the number of plants on that island (2nd column).}
\source{
Sfenthourakis, S. & Triantis K.A. (2009). Habitat diversity,
 ecological requirements of species and the Small Island Effect. Diversity
 Distrib.,15, 131–140.
}
\usage{
data(aegean)
}
\description{
A sample dataset in the correct mmSAR2 format: contains the areas of a number
of islands in the Aegean Sea, Greece, and the number of invertebrate species recorded on each
island.
}
\examples{
data(aegean)
}
\keyword{datasets}
