% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_aegean2.R
\docType{data}
\name{aegean2}
\alias{aegean2}
\title{A SAR dataset describing plants on islands in the Aegean Sea, Greece}
\format{
A data frame with 2 columns and 173 rows. Each row contains the area
  of an island in the Aegean (1st column) and the number of plants on that
  island (2nd column).
}
\source{
Matthews, T.J. et al. (In review) Unravelling the small-island effect
 through phylogenetic community ecology
}
\usage{
data(aegean2)
}
\description{
A sample dataset in the correct sars format: contains the areas of a number
of islands in the Aegean Sea, Greece, and the number of plant species
recorded on each island.
}
\examples{
data(aegean2)
}
\keyword{datasets}
