% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coleman.R
\docType{data}
\name{cole_sim}
\alias{cole_sim}
\title{A simulated species-site abundance matrix with site areas}
\format{A list with two elements. The first element contains a species-site
  abundance matrix in which the rows are species, and the columns are
  sites/islands. Each value in the matrix is the abundance of a species at a
  given site. The second element contains a vector of the areas of each site.}
\source{
Matthews et al. 2015.
}
\usage{
data(cole_sim)
}
\description{
A dataset in the correct sars format:
}
\examples{
data(cole_sim)
}
\keyword{datasets}
