% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_average.R
\name{display_sars_models}
\alias{display_sars_models}
\title{Display the model information table}
\usage{
display_sars_models()
}
\value{
A table of model information for the twenty SAR models, including the
  model function, number of parameters and general model shape.
}
\description{
Display Table 1 of Matthews et al. (2019). See
  \code{\link{sar_multi}} for further information.
}
\references{
Matthews et al. (2019) sars: an R package for fitting, evaluating
  and comparing species–area relationship models. Ecography, In Review.
}
