% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{gdm}
\alias{gdm}
\title{Fit the General Dynamic Model of Island Biogeography}
\usage{
gdm(data, model = "linear", mod_sel = FALSE, A = 1, S = 2, Ti = 3)
}
\arguments{
\item{data}{A dataframe or matrix with at least three columns, where one
column should include island area values, one island richness values and
one island age values.}

\item{model}{Name of the SAR model to be used to fit the GDM. Can be any of
'expo', 'linear', 'power', or 'all'.}

\item{mod_sel}{Logical argument specifying whether, for a given SAR model, a
model comparison of the GDM with other nested candidate models should be
undertaken.}

\item{AST}{The column locations in \code{data} for the area, richness and
time values (in that order).}
}
\value{
An object of class 'gdm'.If \code{model %in% c("expo",
  "linear", "power")} the returned object is a \code{\link{nls}} model fit
  object. If \code{model == "all"}, the returned object is a list with three
  elements; each element being a \code{nls} fit object.

  If \code{mod_sel == TRUE} and \code{model != "all"}, a list with four
  elements is returned; each element being a \code{lm} or \code{nls} fit
  object. When \code{model == "all"}, a list with three elements is returned;
  each element being a list of the four model fits for a particular SAR
  model.
}
\description{
Fit the general dynamic model (GDM) of island biogeography using
  a variety of SAR models. Functions are provided to compare the GDM fitted
  using different SAR models, and also, for a given SAR model, to compare the
  GDM with alternative nested candidate models (e.g. S ~ A + T).
}
\details{
The GDM models island species richness as a function of island area
  and island age, and takes the general form: S ~ A + T + T^2, where S =
  richness, A =area, and T = island age. The T^2 term is included as the GDM
  predicts a hump-shaped relationship between island richness and island age.
  However, a variety of different SAR models have been used to fit the GDM and
  three options are available here: the exponential, linear and power SAR
  model. Model fitting follows the procedure in Cardoso et al. (2015). For
  example, when the linear SAR model is used, the GDM can be fitted using the
  expression: S ~ c + z*Area + k*T + j*T^2, where c,z,k,j are parameters to
  be estimated.

  For all three SAR models, the GDM is fitted using
  non-linear regression and the \code{\link{nls}} function.

  Residual standard error (RSE) is used instead of R2, and for each model the
  AIC and RSE is provided.
  
  null model calculated using lm not nls
  
  linear form of SAR models used
}
\note{
AIC is calculated using the \code{\link{AIC}} function, which is based
  on the log-likelihood and not the residual sum of squares (the latter is
  used in the main functions of the sars package).
  
  A plot generic function enabling 3-d plotting of the GDM fit will be
  provided in a future version of the package.
}
\examples{
#create an example dataset and fit the GDM using the exponential SAR model
data(galap)
galap$t <- rgamma(16, 5, scale = 2)
g <- gdm(galap, model = "expo", mod_sel = FALSE)

#Compare the GDM (using the exponential model) with other nested candidate models
g2 <- gdm(galap, model = "expo", mod_sel = TRUE)

#compare the GDM fitted using the linear, exponential and power SAR models
g3 <- gdm(galap, model = "all", mod_sel = FALSE)
}
\references{
Whittaker, R. J., Triantis, K. A., & Ladle, R. J. (2008). A
  general dynamic theory of oceanic island biogeography. Journal of
  Biogeography, 35, 977-994.

  Borregaard, M. K. et al. (2017). Oceanic island biogeography through the
  lens of the general dynamic model: assessment and prospect. Biological
  Reviews, 92, 830-853.
  
  Cardoso, P., Rigal, F., & Carvalho, J. C. (2015). BAT–Biodiversity
  Assessment Tools, an R package for the measurement and estimation of alpha
  and beta taxon, phylogenetic and functional diversity. Methods in Ecology
  and Evolution, 6, 232-236.
}
