% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_power.R
\name{getLinearPower}
\alias{getLinearPower}
\title{Fit the log-log version of the power model}
\usage{
lin_pow(dat, a = 1, s = 2, con = 1)
}
\arguments{
\item{dat}{A dataset in the form of list. The first element of the list is
the name of the dataset (e.g. "galap"), and the second element is a
dataframe with a minimum of two columns: one with island/site areas, and
one with the species richness of each island/site.}

\item{a}{The column number of the area values. The default is 1 (i.e. column 1)}

\item{s}{The column number of the species richness values. The default is 2 (i.e. column 2)}

\item{con}{The constant to add to the species richness values in cases where one of the islands has zero species}
}
\value{
A list of class 'mmSAR' with a 'Type' attribute
  'lin_pow'. The first element in the list is a vector of overlap values for
  the "in nodes" and the second element is a vector of overlap values for the
  "out nodes".

  The \code{\link{summary.}} methods provides more useful summary
  statistics.
}
\description{
Fit the log-log version of the power model to SAR data and
  return parameter values, summary statistics and the fitted values. A check
  is made for any islands with zero species. If any zero species islands are
  found, a constant (default = 1) is added to each species richness value to
  enable log transformation.Natural logarithms are used.
}
\examples{
data(galap)
y <-  boreal[1:300,] #subset 300 rows for speed
d <- sample(nrow(y), 200, replace = FALSE) #create a random pot_net
pot_net <- y[d,] #by randomly sampling 200 rows from boreal
x <- NOSM_POT_dir(y, pot_net, perc = 1, sl = 1)
summary(x)
}
