% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_habitat.R
\docType{data}
\name{habitat}
\alias{habitat}
\title{A sar_habitat dataset describing snails on islands in the Aegean Sea, Greece}
\format{
A data frame with 3 columns and 12 rows. Each row contains the area
  of an island (1st column), and the number of habitats (2nd column) and
  species on that island (3rd column).
}
\source{
Triantis, K. A., Mylonas, M., Weiser, M. D., Lika, K., &
  Vardinoyannis, K. (2005). Species richness, environmental heterogeneity and
  area: a case study based on land snails in Skyros archipelago (Aegean Sea,
  Greece). Journal of Biogeography, 32, 1727-1735.
}
\usage{
data(habitat)
}
\description{
A sample dataset in the correct sar_habitat format: contains the areas of a
number of islands in the Skyros Archipelago (km2), the number of habitats,
and the number of species recorded on each island.
}
\examples{
data(habitat)
}
\keyword{datasets}
