% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_linear_power.R
\name{lin_pow}
\alias{lin_pow}
\title{Fit the log-log version of the power model}
\usage{
lin_pow(dat, con = 1)
}
\arguments{
\item{con}{The constant to add to the species richness values in cases where
one of the islands has zero species}

\item{x}{A dataset in the form of a dataframe with two columns:
the first with island/site areas, and the second with the species richness of each
island/site.}
}
\value{
A list of class "sars" with four elements. The first element is an
  object of class 'summary.lm'. This is the summary of the linear model fit
  using the \link[stats]{lm} function and the user's data. The second element
  is a numeric vector of the model's fitted values, and the third and fourth
  contain the island areas and observed richness values, respectively.

  The \code{\link{summary.sars}} function returns a more useful summary of the
  model fit results, and the \code{\link{plot}} plots the model.
}
\description{
Fit the log-log version of the power model to SAR data and
  return parameter values, summary statistics and the fitted values. A check
  is made for any islands with zero species. If any zero species islands are
  found, a constant (default = 1) is added to each species richness value to
  enable log transformation. Natural logarithms are used.
}
\examples{
data(galap)
fit <- lin_pow(galap, con = 1)
summary(fit)
plot(fit)
}
