% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niering_data.R
\docType{data}
\name{niering}
\alias{niering}
\title{A SAR dataset describing the plants of the Kaningamarangi Atoll}
\format{A list with two elements. The first element contains the name of the
  dataset. The second element contains a data frame with 2 columns and 16
  rows. Each row contains the area of an island in the Galapagos (1st column)
  and the number of plants on that island (2nd column).}
\source{
Niering, W.A. (1963). Terrestrial ecology of Kapingamarangi Atoll,
 Caroline Islands. Ecol. Monogr., 33, 131–160.
}
\usage{
data(niering)
}
\description{
A sample dataset in the correct mmSAR2 format: contains the areas of a number
of islands in the Kaningamarangi Atoll, and the number of plant species recorded on each
island.
}
\examples{
data(niering)
}
\keyword{datasets}
