% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.mmSAR2}
\alias{plot.mmSAR2}
\title{Plot Model Fits for an mmSAR2 Object}
\usage{
\method{plot}{mmSAR2}(object, ...)
}
\arguments{
\item{object}{An object of class 'mmSAR2'.}

\item{...}{further arguments passed to or from other methods.**see ggplot2 book for examples}
}
\description{
S3 method for class 'mmSAR2'. plot.mmSAR2 creates plots for
  objects of class mmSAR2, using the ggplot2 R package. The exact plot(s)
  constructed depends on the 'Type' attribute (e.g. 'lin_pow') of the mmSAR2
  object.
}
\examples{
data(galap)
fit <- lin_pow(galap, a = 1, s = 2, con = 1)
plot(fit)
}
\references{
Chang.
}
