% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.multi}
\alias{plot.multi}
\title{Plot Model Fits for a 'multi_sars' Object}
\usage{
\method{plot}{multi}(x, type = "multi", allCurves = TRUE, xlab = NULL,
  ylab = NULL, pch = 16, cex = 1.2, pcol = "dodgerblue2",
  ModTitle = NULL, TiAdj = 0, TiLine = 0.5, cex.main = 1.5,
  cex.lab = 1.3, cex.axis = 1, yRange = NULL, pLeg = TRUE, lwd = 2,
  lcol = "dodgerblue2", modNames = NULL, cex.names = 0.88, ...)
}
\description{
Plot Model Fits for a 'multi_sars' Object
}
