% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.sars}
\alias{plot.sars}
\title{Plot Model Fits for an mmsar2 Object}
\usage{
\method{plot}{sars}(x, title = NULL, sh1 = 21, s1 = 3, s2 = 1,
  s3 = 14, s4 = 13, s5 = 15, c1 = "darkred", c2 = "#5e5e5e",
  xl = "Area", yl = "Species richness", p1 = 0, dimen = NULL,
  th = NULL)
}
\arguments{
\item{th}{an additional ggplot2 theme can be added and instructions can be passed.
As ggplot2 is not *loaded*, th needs to use ggplot2::, e.g. ggplot2::theme_bw}

\item{object}{An object of class 'mmSAR2'.}

\item{lsi}{Argument for line width (default = 3).}

\item{ps}{Argument for point size (default = 1).}

\item{pc}{Argument for point colour (default = darkgreen).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method for class 'mmSAR2'. plot.mmSAR2 creates plots for
  objects of class mmSAR2, using the ggplot2 R package. The exact plot(s)
  constructed depends on the 'Type' attribute (e.g. 'lin_pow') of the mmSAR2
  object.

  For an mmSAR2 object of Type 'lin_pow' (i.e. linear power model fit), the
  plot.mmSAR2 function returns a plot of the model fit (black line) and the
  observed richness values (coloured circles).
}
