% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{process_data_statement}
\alias{process_data_statement}
\title{process a data statement}
\usage{
process_data_statement(code)
}
\arguments{
\item{code}{single character string representing data statement}
}
\description{
process a data statement
}
