% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{process_proc_sql_statement}
\alias{process_proc_sql_statement}
\title{process a proc sql statement}
\usage{
process_proc_sql_statement(code)
}
\arguments{
\item{code}{single character string representing proc sql statement}
}
\description{
process a proc sql statement
}
