% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{process_statements}
\alias{process_statements}
\title{Process a bunch of DATA and PROC SQL statements}
\usage{
process_statements(code_vec)
}
\arguments{
\item{code_vec}{vector of character string representing code}
}
\description{
Process a bunch of DATA and PROC SQL statements
}
