% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_asymp.R
\name{sar_asymp}
\alias{sar_asymp}
\title{Fit the Asymptotic regression model}
\usage{
sar_asymp(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}

\item{grid_start}{NULL or the number of points sampled in the model parameter space
or FALSE to prevent any grid start after a fail in inital optimization
to run a grid search.}
}
\description{
Fit the Asymptotic regression model to SAR data
}
\examples{
data(galap)
fit <- sar_asymp(galap)
summary(fit)
plot(fit)
}
