% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_betap.R
\name{sar_betap}
\alias{sar_betap}
\title{Fit the Beta-P cumulative model}
\usage{
sar_betap(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the Beta-P cumulative model to SAR data
}
\examples{
data(galap)
fit <- sar_betap(galap)
summary(fit)
plot(fit)
}
