% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_geom.R
\name{sar_build}
\alias{sar_build}
\title{Build a SAR...}
\usage{
sar_build(data, sp, lurast, crs = 4326,
habNam, habVals, sq_width = 2000, cluster_size_vector)
}
\arguments{
\item{data}{A presence-absence matrix in the form of a
dataframe, where rows are sites. The first three columns
are: site names, then the longitude and latitude of the
sites (note these can be any x and y co-ordinates depending
on the \code{crs} argument provided??). The remaining
columns relate to species in the dataset and take the form
of a standard presence-absence matrix (i.e., 1 for presence
in a given site, 0 for absence).}

\item{sp}{Dataframe of species habitat / specialisation
classifications (e.g., forest specialist, generalist). The
first column must be the species names (which must match the
species column names in \code{data}). The remaining columns
are then the different habitat habitat / specialisation
classifications. Note that these classifications are binary
(i.e., 1 or 0) and a species must only be placed in one
classification, and must have a classification (i.e., the
rows sums of the data frame, minus the first column, must
all equal 1).}

\item{lurast}{A land use raster of class \code{SpatRaster}.
See \code{\link[terra]{SpatRaster}} for more details of this
class.}

\item{crs}{Coordinate reference system, something suitable as
input to \code{\link[sf]{st_crs}}. Default is 'NULL',
whereby the coordinate reference system from \code{lurast}
is used. It is up to the user the ensure the correct
coordinate reference system is employed.}

\item{habNam}{}

\item{habVals}{}

\item{sq_width}{...}

\item{cluster_size_vector}{...}
}
\value{
A dataframe where rows are sites and the first N columns
  are the areas of the different habitats in \code{lurast},
  and then the next n columns are the species richness values
  associated with each species habitat / specialisation
  classification in \code{sp}. These area and richness columns
  can be fed directly into {\link{sar_countryside}}. The last
  two columns provide the total (i.e., summed) area and
  species richness for each site. Note, that these two final
  columns need dropping before using {\link{sar_countryside}}.
}
\description{
Build a SAR .....
}
\details{
***
}
\examples{
data(galap)

}
