% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_chapman.R
\name{sar_chapman}
\alias{sar_chapman}
\title{Fit the Chapman Richards model}
\usage{
sar_chapman(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}

\item{grid_start}{NULL or the number of points sampled in the model parameter space
or FALSE to prevent any grid start after a fail in inital optimization
to run a grid search.}
}
\description{
Fit the Chapman Richards model to SAR data
}
\examples{
data(galap)
fit <- sar_chapman(galap)
summary(fit)
plot(fit)
}
