% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_conf_int.R
\name{sar_conf_int}
\alias{sar_conf_int}
\title{Generate a non parametric confidence interval for a SAR}
\usage{
sar_conf_int(obj, n_boot = 99)
}
\arguments{
\item{obj}{An object of class 'sar' or 'multi.sar'.}

\item{n_boot}{the number of bootstrap sample used to construct the confidence interval.}
}
\description{
Generate a non parametric confidence interval for a SAR
}
\examples{
#multi_sar fit
data(galap)
fit <- multi_sar(data = galap)
conf_fit <- sar_conf_int(fit)
#simple sar fit
data(galap)
fit <- sar_power(data = galap)
conf_fit <- sar_conf_int(fit)
}
