% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_countryside.R
\name{sar_countryside}
\alias{sar_countryside}
\title{Fit the countryside SAR model}
\usage{
sar_countryside(data, modType = "power",
gridStart = "partial", startPar = NULL, zLower = 0, 
ubiSp = FALSE, habNam = NULL, spNam = NULL)
}
\arguments{
\item{data}{A dataset in the form of a dataframe – requires 
a specific column order (see 'Details' below).}

\item{modType}{Fit the power (\code{"power"}) or logarithmic
(\code{"logarithmic"}) form of the countryside model.}

\item{gridStart}{The type of grid search procedure to be
implemented to test multiple starting parameter values: can
be one of \code{partial} (default) or \code{exhaustive}. If
\code{startPar} is provided, this argument is ignored. Note 
that \code{exhaustive} can be quite time consuming to run.}

\item{startPar}{Optional (default = NULL) starting parameter
estimates for the constituent models. Must be a numeric
matrix (see 'Details' below).}

\item{zLower}{The lower bound to be used for the z-parameter
in the \link[minpack.lm]{nlsLM} function. Default is set to
zero, but can be changed to any numeric value (e.g., -Inf to
allow for a full search of parameter space).}

\item{ubiSp}{A logical argument specifying whether a component
model should be fitted for ubiquitous species. If set to
TRUE, a column of ubiquitous species richness must be
included in \code{data}.}

\item{habNam}{Optional vector of habitat names (matching the
column order in \code{data}, otherwise just uses Habitat1
etc).}

\item{spNam}{Optional vector of species-group names (matching
the column order in \code{data}, otherwise takes names of
species columns in \code{data}).}
}
\value{
A list (of class ‘habitat’ and ‘sars’; and with a
  ‘type’ attribute of ‘countryside’) with eight elements: 
  \itemize{ 
   \item \strong{i.}  A list of the non-linear regression model
  fits for each of the species groups.
  \item \strong{ii.}  The habitat affinity values for each of
  the models in (i).
  \item \strong{iii.}   The c-parameter values for each of the
  models in (i).
  \item \strong{iv.}   The predicted total richness values
  (calculated by summing the predictions for each constituent
  countryside model) for each site in the dataset.
  \item \strong{v.}   The residual sum of squares – calculated
  using the predicted and observed total richness values – for
  both the countryside model and the Arrhenius power SAR model
  (or logarithmic model) to enable model comparison.
  \item \strong{vi.}  The dataset used for fitting (i.e., \code{data}).
  \item \strong{vii.}   The power (or logarithmic) model fit object.
  \item \strong{viii.}   The user-provided \code{ubiSp} argument.}
}
\description{
Fit the countryside biogeography SAR model in
  either power or logarithmic form, including an optional
  component model for ubiquitous species.
}
\details{
To work, the countryside SAR model requires that all
  species in the study system have been classified based on
  the habitats present. For example, in a study system with
  two habitats (forest and grassland), all species must a
  priori have been classified as either forest species or
  grassland species; optionally, species can also be
  classified as ubiquitous species (i.e., species that do not
  have strong affinity for a specific habitat – true habitat
  generalists; controlled using the \code{ubiSp} argument).
  The provided input dataset (\code{data}) will typically
  relate to a series of landscapes with differing areas of the
  N habitats (e.g., forest and grassland), and for each
  landscape the number of (for example) forest species and
  grassland species present will be provided as well as
  (optionally) the number of ubiquitous species.


##COLUMN ORDER NO LONGER IMPORTANt but habitat cols must be first,
then species columns. In the model output, the h coefficients follow 
the order of the habitat cols in data (e.g., h1 = col 1)

It is important that the column orders in \code{data} are
correct. The first set of columns should be the habitat area
columns, followed by the habitat species richness columns
(these should be in the same order as the area columns). An
optional final column of ubiquitous species richness can also
be included. For example, in a
dataset with two habitats (forest and grassland) and setting \code{ubiSp =
TRUE}, the column order in \code{data} could be:
forest-a, grassland-a, forest-s, grassland-s, and Ubi-s, where
a = area, s = species richness, and Ubi-s = the number of
ubiquitous species.

The countryside SAR model works by fitting individual
component models of a particular form (e.g., power), one for
each of the habitat types (e.g., one model for forest species,
one for grassland species, and so on). The predictions from
these component models are then combined to generate a total
predicted richness for each site / landscape. The
output of the model fitting includes the individual component
model fits, the total predicted (fitted) richness values for
each site, and the habitat affinity values for each species
group. The latter vary from zero to one, and equal 1 for a
given species group's affinity to its preferred habitat (e.g.,
forest species for forest).
  
For \code{startPar}, if not NULL, it needs to be a numeric
matrix, where number of rows = number of species groups
(including ubiquitous sp., if provided), and number of columns
equals number of habitats + 1. Matrix row order matches the
order of species group columns in \code{data}, and matrix
column order matches the order of habitat columns in
\code{data} + 1 extra final column for the z-parameter
estimate.

Two different types of plot can be generated with the output,
using \code{\link{plot.habitat}}. The
\code{\link{countryside_extrap}} function can be used with the
output of \code{sar_countryside} to predict the species
richness of landscapes with varying areas of the analysed
habitats.

See Matthews et al. (2025) for further details.
}
\note{
The model fits in (i) are objects of class ‘nls’,
  meaning that all the basic non-linear regression R methods
  can be applied (e.g., generating model summary tables or
  plotting the model residuals).
}
\examples{
data(countryside)
\dontrun{
#Fit the countryside SAR model (power form) to the data.
#Include a component model of ubiquitous species, and use the
#function’s starting parameter value selection procedure.
#Abbreviations: AG = agricultural land, SH = shrubland, F =
#oak forest, UB = ubiquitous species.
s3 <- sar_countryside(data = countryside, modType = "power",
gridStart = "partial", ubiSp = TRUE, habNam = c("AG", "SH",
"F"), spNam = c("AG_Sp", "SH_Sp", "F_Sp", "UB_Sp"))

#Predict the richness of a site which comprises 1000 area units
#of agricultural land, 1000 of shrubland and 1000 of forest.
countryside_extrap(s3, area = c(1000, 1000, 1000))

#Generate a plot of the countryside model’s predicted total
#richness vs. the observed total richness, and include the
#predictions of the Arrhenius power model

plot(s3, type = 1, powFit = TRUE)

#Plot the fitted individual SAR curves for each species group,
#providing set line colours, including a legend and
#positioning it outside the main plotting window, and modifying
#other aspects of the plot using the standard base R plotting
#commands
par(mar=c(5.1, 4.1, 4.1, 7.5), xpd=TRUE)
plot(s3, type = 2, lcol = c("black", "aquamarine4",
"#CC661AB3" , "darkblue"), pLeg = TRUE,
legPos ="topright", legInset = c(-0.27,0.3), lwd = 1.5)

#Provide starting parameter estimates for the component models
#instead of using gridStart
M2 <- matrix(c(3.061e+08, 2.105e-01, 1.075e+00, 1.224e-01,
3.354e-08, 5.770e+05, 1.225e+01, 1.090e-01,
6.848e-01, 1.054e-01, 4.628e+05, 1.378e-01,
0.20747, 0.05259, 0.49393, 0.18725), nrow = 4,
byrow = TRUE)

s4 <- sar_countryside(data = countryside,
                    modType = "power",
                   startPar = M2, ubiSp = TRUE)
}
}
\references{
Matthews et al. (2025) In prep.

Pereira, H.M. & Daily, G.C. (2006) Modelling biodiversity
dynamics in countryside landscapes. Ecology, 87, 1877–1885.
  
  Proença, V. & Pereira, H.M. (2013) Species–area models to
  assess biodiversity change in multi-habitat landscapes: the
  importance of species habitat affinity. Basic and Applied
  Ecology, 14, 102–114.
}
\author{
Thomas J. Matthews, Inês Santos Martins, Vânia Proença 
and Henrique Pereira
}
