% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_epm2.R
\name{sar_epm2}
\alias{sar_epm2}
\title{Fit the Extended Power model 2 model}
\usage{
sar_epm2(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the Extended Power model 2 model to SAR data
}
\examples{
data(galap)
fit <- sar_epm2(galap)
summary(fit)
plot(fit)
}
