% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_habitat.R
\name{sar_habitat}
\alias{sar_habitat}
\title{Fit habitat SAR models}
\usage{
sar_habitat(data, logAxes = "both", con = 1, logT = log)
}
\arguments{
\item{data}{A dataset in the form of a dataframe with at least ** columns:
the first with island/site areas, and the second with the species richness
of each island/site.}

\item{logAxes}{What log-transformation should be applied to the area and
richness values. Should be one of  "area" (only area is log-transformed;
default) or "both" (both area and richness log-transformed).}

\item{con}{The constant to add to the species richness values in cases where
one of the islands has zero species.}

\item{logT}{The log-transformation to apply to the area and richness values.
Can be any of \code{log}(default), \code{log2} or \code{log10}.}
}
\value{
A list of class "threshold" and "sars" with five elements. The first
  element contains the different model fits (lm objects). The second element
  contains the names of the fitted models, the third  contains the threshold
  values, the fourth element the dataset (i.e. a dataframe with area and
  richness values), and the fifth contains details of any axes
  log-transformations undertaken. \code{\link{summary.sars}} provides a more
  user-friendly ouput (including a model summary table) and
  \code{\link{plot.threshold}} plots the model fits.
}
\description{
Fit three SAR regression models including habitat divesity data.
}
\details{
This function is described in more detail in the accompanying paper
  (Matthews & Rigal, 2020).

  Fitting the continuous and left-horizontal piecewise models (particularly
  the two-threshold models) can be time consuming if the range in area is
  large and/or the \code{interval} argument is small. For the two-threshold
  continuous slope and left-horizontal models, the use of parallel processing
  (using the \code{parallel} argument) is recommended. The number of cores
  (\code{cores}) must be provided.
}
\note{
Due to the
}
\examples{
data(aegean2)
a2 <- aegean2[1:168,]
fitT <- sar_threshold(data = a2, mod = c("ContOne", "DiscOne"), 
interval = 0.1, non_th_models = TRUE, logAxes = "area", logT = log10)
summary(fitT)
plot(fitT)
#diagnostic plots for the ContOne model
par(mfrow=c(2, 2))
plot(fitT[[1]][[1]])
}
\references{
Furness, E.N., Saupe, E.E., Garwood, R.J., Mannion, P.D. &
  Sutton, M.D. (2023) The jigsaw model: a biogeographic model that partitions
  habitat heterogeneity from area. Frontiers of Biogeography, 15, e58477.

  Triantis, K.A., Mylonas, M., Lika, K. & Vardinoyannis, K. (2003) A model
  for the species-area-habitat relationship. Journal of Biogeography, 30,
  19–27.
}
\author{
Euan N. Furness and Thomas J. Matthews
}
