% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_heleg.R
\name{sar_heleg}
\alias{sar_heleg}
\title{Fit the Logistic (He & Legendre) model}
\usage{
sar_heleg(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}

\item{grid_start}{NULL or the number of points sampled in the model parameter space
or FALSE to prevent any grid start after a fail in inital optimization
to run a grid search.}
}
\description{
Fit the Logistic (He & Legendre) model to SAR data
}
\examples{
data(galap)
fit <- sar_heleg(galap)
summary(fit)
plot(fit)
}
