% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_linear.R
\name{sar_linear}
\alias{sar_linear}
\title{Fit the linear model}
\usage{
sar_linear(data,)
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the linear model to SAR data
}
\examples{
data(galap)
fit <- sar_linear(galap)
summary(fit)
plot(fit)
}
