% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_monod.R
\name{sar_monod}
\alias{sar_monod}
\title{Fit the Monod model}
\usage{
sar_monod(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the Monod model to SAR data
}
\examples{
data(galap)
fit <- sar_monod(galap)
summary(fit)
plot(fit)
}
