% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_powerR.R
\name{sar_powerR}
\alias{sar_powerR}
\title{Fit the PowerR model}
\usage{
sar_powerR(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the PowerR model to SAR data
}
\examples{
data(galap)
fit <- sar_powerR(galap)
summary(fit)
plot(fit)
}
