% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmSAR2.R
\docType{package}
\name{sars-package}
\alias{sars-package}
\title{sars: Fit and compare species-area relationship models using multimodel inference}
\description{
A package to fit twenty species-area relationship models . ..
}
\details{
There are ***
}
\examples{
data(galap, package = "sars")
fit = (galap)
summary(fit)
plot(fit)
}
\references{
Triantis, K.A., Guilhaumon, F. & Whittaker, R.J. (2012) The
  island species–area relationship: biology and statistics. Journal of
  Biogeography, 39, 215-231.

  Matthews, T.J., Guilhaumon, F., Triantis, K.A, Borregaard, M.K., &
  Whittaker, R.J. (2015b) On the form of species–area relationships in
  habitat islands and true islands. Global Ecology & Biogeography. DOI:
  10.1111/geb.12269.
}
\seealso{
\url{https://github.com/txm676/sars}
}
\author{
Thomas J. Matthews and Francois Guilhaumon
}
