% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{split_into_statements}
\alias{split_into_statements}
\title{split code into data and proc statements}
\usage{
split_into_statements(code)
}
\arguments{
\item{code}{string containing code}
}
\value{
character vector of statements
}
\description{
split code into data and proc statements
}
