% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_summary.R
\name{summary.sars}
\alias{summary.sars}
\title{Summarising the results of the model fitting functions}
\usage{
\method{summary}{sars}(object)
}
\arguments{
\item{object}{An object of class 'sars'.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The summary function returns an object of class "summary.sars". A
  print function is used to obtain and print a summary of the model fit
  results.
}
\description{
S3 method for class 'sars'. summary.sars creates summary
  statistics for objects of class 'sars'. The exact summary statistics
  computed depends on the 'Type' attribute (e.g. 'lin_pow') of the 'sars'
  object. The summary method generates more useful information
  for the user than the standard model fitting functions. Another S3 method
  (\code{print.summary.sars}; not documented) is used to print the output.
}
\examples{
data(galap)
fit <- lin_pow(galap, con = 1, compare = T)
summary(fit)
}
