% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Table1.R
\docType{data}
\name{Table1}
\alias{Table1}
\title{Model information table}
\format{A data frame with 4 columns and 20 rows (corresponding to the 20 SAR models).}
\source{
Matthews et al. (2019) sars: an R package for fitting, evaluating
  and comparing species–area relationship models. Ecography, In Review.
}
\usage{
data(Table1)
}
\description{
A table of model information for the twenty SAR models, including the
  model function, number of parameters and general model shape.
}
\examples{
data(Table1)
}
\keyword{datasets}
