% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_average.R
\name{display_sars_models}
\alias{display_sars_models}
\title{Display the model information table}
\usage{
display_sars_models()
}
\value{
A table of model information for 21 SAR models, including the model
  function, number of parameters and general model shape. This includes the
  20 models in Matthews et al. (2019); however, note that the mmf model has
  now been deprecated, and the standard logistic model listed in Tjorve
  (2003) added instead. Note also, an error in the Chapman Richards model
  equation has now been corrected, and the shape of some of the models have
  been updated from sigmoid to convex/sigmoid.
}
\description{
Display Table 1 of Matthews et al. (2019). See
  \code{\link{sar_multi}} for further information.
}
\references{
Matthews et al. (2019) sars: an R package for fitting, evaluating
  and comparing species–area relationship models. Ecography,  42, 1446-1455.

  Tjørve, E. (2003) Shapes and functions of species–area curves: a review of
  possible models. Journal of Biogeography, 30, 827-835.
}
