% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdm.R
\name{gdm}
\alias{gdm}
\title{Fit the General Dynamic Model of Island Biogeography}
\usage{
gdm(data, model = "lin_pow", mod_sel = FALSE, A = 1, S = 2, Ti = 3)
}
\arguments{
\item{data}{A dataframe or matrix in which rows are species and columns are
sites. Each element/value in the matrix is the abundance of a given species
in a given site.}

\item{area}{A vector of site (island) area values. The order of the vector
must match the order of the columns in \code{data}.}
}
\value{
A list of class "coleman" with four elements. The first element
  contains the fitted values of the model. The second element contains the
  standard deviations of the fitted values, and the third and fourth contain
  the relative island areas and observed richness values, respectively.
  \code{\link{plot.coleman}} plots the model.
}
\description{
Fit Coleman's (1981) random placement model to a species-site
  abundance matrix: rows are species and columns are sites. Note that the
  data must be abundance data and not presence-absence data. According to
  this model, the number of species occurring on an island depends on the
  relative area of the island and the regional relative species abundances.
  The fit of the random placement model can be determined through use of a
  diagnostic plot (see \code{\link{plot.coleman}}) of island area (log
  transformed) against species richness, alongside the model’s predicted
  values (see Wang et al., 2010). Following Wang et al. (2010), the model is
  rejected if more than a third of the observed data points fall beyond one
  standard deviation from the expected curve.
}
\examples{
#create an example dataset and fit the GDM using the exponential SAR model
data(galap)
galap$t <- rgamma(16, 5, scale = 2)
g <- gdm(galap, model = "expo", mod_sel = FALSE)

#Compare the GDM (using the exponential model) with other **
g2 <- gdm(galap, model = "expo", mod_sel = TRUE)

#compare the GDM fitted using the linear, exponential and power SAR models
g3 <- gdm(galap, model = "all", mod_sel = FALSE)
}
\references{
Whittaker, R. J., Triantis, K. A., & Ladle, R. J. (2008). A
  general dynamic theory of oceanic island biogeography. Journal of
  Biogeography, 35, 977-994.
  
  Borregaard, M. K. et al. (2017). Oceanic island biogeography through the
  lens of the general dynamic model: assessment and prospect. Biological
  Reviews, 92, 830-853.
}
