% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_niering.R
\docType{data}
\name{niering}
\alias{niering}
\title{A SAR dataset describing the plants of the Kapingamarangi Atoll}
\format{A data frame with 2 columns and 32 rows. Each row contains the area
  of an island (km2) in the Kapingamarangi Atoll (1st column) and the number
  of plants on that island (2nd column).}
\source{
Niering, W.A. (1963). Terrestrial ecology of Kapingamarangi Atoll,
 Caroline Islands. Ecol. Monogr., 33, 131–160.
}
\usage{
data(niering)
}
\description{
A sample dataset in the correct sars format: contains the areas of a number
of islands in the Kapingamarangi Atoll, and the number of plant species
recorded on each island.
}
\examples{
data(niering)
}
\keyword{datasets}
