% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.habitat}
\alias{plot.habitat}
\title{Barplot of information criteria weights for a 'habitat' Object}
\usage{
\method{plot}{habitat}(
  x,
  IC = "AICc",
  type = 1,
  powFit = TRUE,
  lcol = NULL,
  pLeg = TRUE,
  legPos = "bottomright",
  legInset = 0,
  ...
)
}
\arguments{
\item{x}{An object of class 'habitat'.}

\item{IC}{The information criterion weights to present (must be one of 'AIC',
'BIC' or 'AICc')}

\item{\dots}{Further graphical parameters may be supplied as arguments.}
}
\description{
S3 method for class 'habitat'. \code{plot.habitat} creates
 a simple barplot of information criteria weights for the different model fits
}
\examples{
#data(habitat)
###TO FILL IN
##legPos = can either be position (e.g., "bottomright"), 
##or the x and y co-ordinates to be used to position the legend,
##(eg., c(0,5))
##legInset = the inset argument in graphics::legend - lets you
##plot the legend outside the plotting window (needs the user to
##to manually change their graphics pars)
}
