% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.habitat}
\alias{plot.habitat}
\title{Barplot of information criteria weights for a 'habitat' Object}
\usage{
\method{plot}{habitat}(x, IC = "AICc", ...)
}
\arguments{
\item{x}{An object of class 'habitat'.}

\item{IC}{The information criterion weights to present (must be one of 'AIC',
'BIC' or 'AICc')}

\item{\dots}{Further graphical parameters may be supplied as arguments.}
}
\description{
S3 method for class 'habitat'. \code{plot.habitat} creates
 a simple barplot of information criteria weights for the different model fits
}
\examples{
#data(habitat)
###TO FILL IN
}
