% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_plot.R
\name{plot.mmsar2}
\alias{plot.mmsar2}
\title{Plot Model Fits for an mmsar2 Object}
\usage{
\method{plot}{mmsar2}(x, title = NULL, sh1 = 16, s1 = 2, s2 = 1,
  s3 = 14, s4 = 13, s5 = 15, c1 = "darkred", c2 = "black",
  xl = "Area", yl = "Species richness", p1 = 0)
}
\arguments{
\item{object}{An object of class 'mmSAR2'.}

\item{lsi}{Argument for line width (default = 3).}

\item{ps}{Argument for point size (default = 1).}

\item{pc}{Argument for point colour (default = darkgreen).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method for class 'mmSAR2'. plot.mmSAR2 creates plots for
  objects of class mmSAR2, using the ggplot2 R package. The exact plot(s)
  constructed depends on the 'Type' attribute (e.g. 'lin_pow') of the mmSAR2
  object.

  For an mmSAR2 object of Type 'lin_pow' (i.e. linear power model fit), the
  plot.mmSAR2 function returns a plot of the model fit (black line) and the
  observed richness values (coloured circles).
}
