% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_countryside.R
\name{sar_countryside}
\alias{sar_countryside}
\title{Fit the countryside SAR model}
\usage{
sar_countryside(
  data,
  modType = "power",
  gridStart = "partial",
  startPar = NULL,
  zLower = 0,
  ubiSp = FALSE,
  spNam = NULL,
  habNam = NULL
)
}
\arguments{
\item{data}{An object of class 'habitat'.}

\item{modType}{The information criterion weights to present (must be one of 'AIC',}

\item{gridStart}{description}

\item{startPar}{description}

\item{zLower}{de}

\item{ubiSp}{description}

\item{spNam}{description}

\item{habNam}{description}
}
\description{
A short description...
}
