% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_epm1.R
\name{sar_epm1}
\alias{sar_epm1}
\title{Fit the Extended Power model 1 model}
\usage{
sar_epm1(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the Extended Power model 1 model to SAR data
}
\examples{
data(galap)
fit <- sar_epm1(galap)
summary(fit)
plot(fit)
}
