% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_habitat.R
\name{sar_habitat}
\alias{sar_habitat}
\title{Fit habitat SAR models}
\usage{
sar_habitat(data, modType = "power_log", con = 1, logT = log)
}
\arguments{
\item{data}{A dataset in the form of a dataframe with at least three columns:
the first with island/site areas, the second with island / site habitat
diversity, and the third with the species richness of each island/site.}

\item{modType}{What underlying SAR model form should be used. Should be one
of "power" (non-linear power), "logarithmic" (logarithmic SAR), or
"power_log" (log-log power; default).}

\item{con}{The constant to add to the species richness values in cases where
at least one of the islands has zero species.}

\item{logT}{The log-transformation to apply to the area and richness values.
Can be any of \code{log}(default), \code{log2} or \code{log10}.}
}
\value{
A list of class "habitat" and "sars" with four elements, each holding
  one of the individual model fit objects (either ** or lm objects).
  \code{\link{summary.sars}} provides a more user-friendly ouput (including a
  model summary table ranked by AICc and presenting the model coefficients,
  and R2 and information criteria values etc.) and \code{\link{plot.habitat}}
  provides a simple bar of information criteria weights. For the models
  fitted using non-linear regression, the R2 and adjusted R2 are 'pseudo R2'
  values and are calculated using the same approach as in the
  \code{\link{sar_average}} function.
}
\description{
Fit three SAR regression models that include habitat diversity:
the choros model, the Kallimanis model, and the jigsaw model.
}
\details{
These functions are described in more detail in the accompanying paper
  (Furness et al., 2023). The code to fit the models was also taken from this
  paper.
  
  Three habitat SAR models are available:
  \itemize{ \item{choros model:} { Proposes that species richness is better 
  predicted by the product of habitat heterogeneity and area (S = c(AH)^z) } 
  \item{Kallimanis model:} { 
  Proposes that increasing habitat heterogeneity increases species richness 
  by increasing the slope (on a log-log plot) of the Arrhenius model } 
  \item{jigsaw model:} {
  Models species richness in an area as the sum of the species richness 
  values of several smaller component subareas, which can be visualised as 
  pieces of a jigsaw puzzle, i.e., it partitions the species–area and 
  species–heterogeneity scaling relationships }}
  
  In addition to these three models, a simple 'non-habitat' SAR model is also
  fit, which varies depending on \code{modType}: the non-linear power, the
  logarithmic or the log-log power model.
}
\note{
The jigsaw model is equivalent to the trivariate power-law model of 
Tjørve (2009), see Furness et al. (2023).

The jigsaw model (power-law form) cannot have a poorer fit than the choros or
power model based on RSS and thus R2. Comparing models using information
criteria is thus advised.
}
\examples{
data(habitat)
#Fit the models in log-log space
s <- sar_habitat(data = habitat, modType = "power_log", 
con = 1, logT = log)
#Look at the model comparison summary
s2 <- summary(s)
s2
#Make a simple plot of AICc weights
plot(s, IC = "AICc", col = "darkred")
}
\references{
Furness, E.N., Saupe, E.E., Garwood, R.J., Mannion, P.D. &
  Sutton, M.D. (2023) The jigsaw model: a biogeographic model that partitions
  habitat heterogeneity from area. Frontiers of Biogeography, 15, e58477.
  
  Kallimanis, A.S., Mazaris, A.D., Tzanopoulos, J., Halley, J.M., Pantis,
  J.D., & Sgardelis, S.P. (2008) How does habitat diversity affect the
  species–area relationship? Global Ecology and Biogeography, 17, 532-538
  
  Tjørve, E. (2009) Shapes and functions of species– area curves (II): a
  review of new models and parameterizations. Journal of Biogeography, 36,
  1435-1445.
  
  Triantis, K.A., Mylonas, M., Lika, K. & Vardinoyannis, K. (2003) A model
  for the species-area-habitat relationship. Journal of Biogeography, 30,
  19–27.
}
\author{
Euan N. Furness and Thomas J. Matthews
}
