% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_heleg.R
\name{sar_heleg}
\alias{sar_heleg}
\title{Fit the Asymptotic regression model}
\usage{
sar_heleg(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}
}
\description{
Fit the Asymptotic regression model to SAR data
}
\examples{
data(galap)
fit <- sar_heleg(galap)
summary(fit)
plot(fit)
}
