% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_koba.R
\name{sar_koba}
\alias{sar_koba}
\title{Fit the Kobayashi model}
\usage{
sar_koba(data, custstart = NULL, normtest = 'lillie')
}
\arguments{
\item{data}{A dataset in the form of a dataframe with two columns: 
the first with island/site areas, and the second with the species richness
of each island/site.}

\item{grid_start}{NULL or the number of points sampled in the model parameter space
or FALSE to prevent any grid start after a fail in inital optimization
to run a grid search.}
}
\description{
Fit the Kobayashi model to SAR data
}
\examples{
data(galap)
fit <- sar_koba(galap)
summary(fit)
plot(fit)
}
