% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_average.R
\name{sars_models}
\alias{sars_models}
\title{Display the 21 SAR model names}
\usage{
sars_models()
}
\value{
A vector of model names.
}
\description{
Display the 21 SAR model names as a vector. See
  \code{\link{sar_multi}} for further information.
}
\note{
sar_mmf is included here for now but has been deprecated (see News)
}
