% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table1.R
\docType{data}
\name{Table1}
\alias{Table1}
\title{List of Species Areas Relationships included.}
\format{A data frame,one row per SAR.}
\usage{
data(Table1)
}
\description{
For each of the SARs included, the table includes the name of the SAR,
the name of the corresponding function, the number of parameters, the
function name, the number of parameters, the equation and the shape of the
model.
}
\examples{
data(Table1)
}
\keyword{datasets}
