% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_threshold.R
\name{threshold_ci}
\alias{threshold_ci}
\title{Calculate confidence intervals around breakpoints}
\usage{
threshold_ci(object, interval = NULL, Nboot = 100, verb = TRUE)
}
\arguments{
\item{object}{An object of class 'thresholds', generated using the
\code{\link{sar_threshold}} function. The object must contain fits of
either (or both) of the one-threshold continuous or the one-threshold
left-horizontal model.}

\item{interval}{The amount to increment the threshold value by in the
iterative model fitting process. The default for non-transformed area
reverts to 1, while for log-transformed area it is 0.01. It is advised that
the same interval value used when running \code{\link{sar_threshold}}
is used here.}

\item{Nboot}{Number of bootstrap samples.}

\item{verb}{Should progress be reported. If \code{TRUE}, every 50th bootstrap
sample is reported.}
}
\value{
A list of class "sars" with two elements. The first element contains
  the full set of bootstrapped breakpoint estimates for each model. The
  second contains the 95 percent confidence interval values.
}
\description{
Generate confidence intervals around the breakpoints of the
  one-threshold continuous and left-horizontal models, using bootstrapping.
}
\details{
Calculates confidence intervals (95 percent) around the breakpoint
  estimates using the bootstrap approach of Toms & Lesperance (2003). If the
  number of bootstrap samples (\code{Nboot}) is large, the function can take
  a while to run.

  Currently only available for the one-threshold continuous and left-
  horizontal threshold models.
}
\examples{
data(aegean2)
a2 <- aegean2[1:168,]
fitT <- sar_threshold(data = a2, mod = c("ContOne", "DiscOne"), 
interval = 0.1, non_th_models = TRUE, logAxes = "area", logT = log10)
#calculate confidence intervals (using very low Nboot just as an example)
CI <- threshold_ci(fitT, interval = NULL, Nboot = 3)
CI
}
\references{
Toms, J.D. & Lesperance, M.L. (2003) Piecewise regression: a tool
  for identifying ecological thresholds. Ecology, 84, 2034-2041.
}
